refactor: split report generators into modular package
